requirements.txt
beautifulsoup4==4.12.3
Faker==27.4.0
Flask==3.0.3
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
flask_wtf==1.2.1
MarkupSafe==2.1.3
python-slugify==8.0.4
Requests==2.32.3
SQLAlchemy==2.0.25
stripe==9.5.0
Werkzeug==3.0.4
WTForms==3.1.2